requirements.txt
Django>=4.2
djangorestframework
pandas